refactor: separate `tvm_model_t` from embedding_model to tvm_model
